requirements.txt
streamlit
google-cloud-bigquery
anthropic
requests